🚚 feat : moved interface to model
