+++ .gitignore
@@ -1,6 +1,9 @@
 # Frontend 
 frontend/node_modules
 frontend/coverage
+frontend/tsconfig.app.tsbuildinfo
+frontend/tsconfig.node.tsbuildinfo
+
 
 # Logs
 logs
