+++ frontend/sonar-project.properties
@@ -3,7 +3,7 @@ sonar.organization=dev4beginners-gh
 sonar.host.url=https://sonarcloud.io
 sonar.projectVersion=1.0
 sonar.sources=src
-sonar.exclusions=**/node_modules/**, **/*.stories.tsx , **/src/theme/**, **/src/**/styled.ts, **/src/utils/**, **/src/services/**,**/src/context/**,**/src/routes/**
+sonar.exclusions=**/node_modules/**, **/*.stories.tsx , **/src/theme/**, **/src/**/styled.ts, **/src/utils/**, **/src/services/**,**/src/context/**,**/src/routes/**, **/src/models/**
 sonar.tests=src
 sonar.test.inclusions=**/*.test.ts,**/*.test.tsx
 sonar.test.exclusions=**/*.stories.tsx
