+++ frontend/src/models/interfaces.ts
